chnages after revertng back to prisma v6

+++ lib/actions/user.actions.ts
@@ -16,7 +16,7 @@ import { ShippingAddress } from "@/types";
 import { z } from "zod";
 import { PAGE_SIZE } from "../constants";
 import { revalidatePath } from "next/cache";
-import { Prisma } from "@/db/generated/prisma/client.ts";
+import { Prisma } from "@prisma/client";
 import { getMyCart } from "./cart.actions";
 
 //sign in the user with credentials
@@ -45,8 +45,8 @@ export async function signInWithCredentials(
 export async function signOutUser() {
   // get current users cart and delete it so it does not persist to next user
   const currentCart = await getMyCart();
-  await prisma.cart.delete({ where: { id: currentCart?.id } });
-  await signOut({ redirectTo: "/" });
+  if (currentCart) await prisma.cart.delete({ where: { id: currentCart?.id } });
+  await signOut();
 }
 
 //sign up user
